% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTextGrid.R
\name{createTextGrid}
\alias{createTextGrid}
\title{Create a new TextGrid object}
\usage{
createTextGrid(xmin, xmax, tier_names, point_tiers)
}
\arguments{
\item{xmin}{numeric, minimum time value}

\item{xmax}{numberic, maximum time value}

\item{tier_names}{character vector of tier names}

\item{point_tiers}{subset character vector of tier names that are to be point tiers, or NA if none}
}
\description{
Create a new TextGrid in R
}
\examples{
# Create the default Praat TextGrid
createTextGrid(0, 1, c("Mary", "John", "bell"), "bell")

}

