% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTierByName.R
\name{getTierByName}
\alias{getTierByName}
\title{Select TextGrid Tier by its name}
\usage{
getTierByName(tg, tier_name)
}
\arguments{
\item{tg}{TextGrid data, returned from readTextGrid() function}

\item{tier_name}{character string. The name of the tier to retrive from TextGrid \code{tg}}
}
\description{
This function takes a TextGrid and a name (character string), and returns the specified tier from that TextGrid
}
\examples{
# Get word tier from data in mytg.TextGrid
tg_data <- readTextGrid("mytg.TextGrid")
getTierByName(tg_data, "word")

# You may also chain with \%>\% from dplyr
readTextGrid("mytg.TextGrid") \%>\%
   getTierByName("word")

}

