% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTierByNumber.R
\name{getTierByNumber}
\alias{getTierByNumber}
\title{Select TextGrid Tier by its number}
\usage{
getTierByNumber(tg, tier_number)
}
\arguments{
\item{tg}{TextGrid data, returned from readTextGrid() function}

\item{tier_number}{integer. A number of the tier to retrive from TextGrid \code{tg}}
}
\description{
This function takes a TextGrid and a number, and returns the specified tier from that TextGrid
}
\examples{
# Get tier number 1 from data in mytg.TextGrid
tg_data <- readTextGrid("mytg.TextGrid")
getTierByNumber(tg_data, 1)

# You may also chain with \%>\% from dplyr
readTextGrid("mytg.TextGrid") \%>\%
   getTierByNumber(1)
}

