% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTierIntervals.R
\name{getTierIntervals}
\alias{getTierIntervals}
\title{Retrieve all Intervals from an Interval Tier}
\usage{
getTierIntervals(interval_tier, discard_empty = FALSE)
}
\arguments{
\item{interval_tier}{an IntervalTier returned from getTierByName() or getTierByNumber()}

\item{discard_empty}{boolean FALSE (default) or TRUE. Indicates whether intervals with no text in them should be discarded in selection.}
}
\description{
This function returns a data frame of the Intervals from an Interval Tier.
}
\examples{
# Get intervals from "word" tier from data in mytg.TextGrid
tg_data <- readTextGrid("mytg.TextGrid")
word_tier <- getTierByName(tg_data, "word")
word_intervals <- getTierIntervals(word_tier)

# You may also chain with \%>\% from dplyr
readTextGrid("mytg.TextGrid") \%>\%
   getTierByName("word") \%>\%
   getTierIntervals()
}

