% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTextGrid.R
\name{readTextGrid}
\alias{readTextGrid}
\title{Read in data from a Praat TextGrid file}
\usage{
readTextGrid(file, format = "R")
}
\arguments{
\item{file}{a character string of the filename to read}

\item{format}{optional. A character string "R" (default) or "JSON" of the data format to be returned.}
}
\description{
Read data from a TextGrid file into R
}
\examples{
readTextGrid("my_tg.TextGrid")
readTextGrid("path/to/textgrids/thisfile.TextGrid")
}

