% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTextGridDirectory.R
\name{readTextGridDirectory}
\alias{readTextGridDirectory}
\title{Read in data from all Praat TextGrid files within a directory}
\usage{
readTextGridDirectory(dir = ".", sources = TRUE, ...)
}
\arguments{
\item{dir}{Optional. The directory to read in. Default is the current working directory.}

\item{sources}{boolean TRUE (default) or FALSE - whether to append a 'source' column to each TextGrid data with its filename}
}
\description{
The function finds all .TextGrid or .textgrid (case insensitive) files within a directory and returns a list of the TextGrid data
}
\examples{
readTextGridDirectory("/path/to/dir")

# Read all TextGrid files in current working directory
readTextGridDirectory()
}

