% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTextGrid.R
\name{writeTextGrid}
\alias{writeTextGrid}
\title{Write data to a Praat TextGrid file}
\usage{
writeTextGrid(tg, file)
}
\arguments{
\item{tg}{TextGrid data, returned from readTextGrid() or createTextGrid() functions}

\item{file}{a character string of the filename to write to}
}
\description{
Write data to a TextGrid file into R
}
\examples{
Write to the current working directory
writeTextGrid(tg_data, "my_tg.TextGrid")

Write to specific file path
readTextGrid(tg_data, "path/to/textgrids/thisfile.TextGrid")
}

